Fix int => float64

+++ docker-structs.go
@@ -24,13 +24,13 @@ type DockerInfo struct {
 }
 
 type Container struct {
-	ID         string `json:"Id"`
-	Hostname   string `json:"Hostname"`
-	Image      string `json:"Image"`
-	IPAddress  string `json:"IPAddress"`
-	MacAddress string `json:"MacAddress"`
-	SizeRootFs int    `json:"SizeRootFs"`
-	SizeRw     int    `json:"SizeRw"`
+	ID         string  `json:"Id"`
+	Hostname   string  `json:"Hostname"`
+	Image      string  `json:"Image"`
+	IPAddress  string  `json:"IPAddress"`
+	MacAddress string  `json:"MacAddress"`
+	SizeRootFs float64 `json:"SizeRootFs"`
+	SizeRw     float64 `json:"SizeRw"`
 	State      struct {
 		Dead       bool   `json:"Dead"`
 		Error      string `json:"Error"`
